// src/app/interfaces/group.interface.ts
export interface GroupData {
    group_name: string;
    BFSCIT_TIME: any;
    BFSCOT_TIME: any;
    CASCIT_AM_TIME: any;
    CASCOT_AM_TIME: any;
    CASCIT_PM_TIME: any;
    CASCOT_PM_TIME: any;
    AFSCIT_TIME: any;
    AFSCOT_TIME: any;
    create_at: number;
    update_at: number;
    status: number;
    recurrentAm: any;
    recurrentPm: any;
    vacations: any;
    agent: string;
}